convert from bytes for latin1 missing text in JS
When jsQR is used, there might be a missing "text" entry
altough there exists a "bytes" entry in the data returned.
The missing "text" entry is probably due to the UTF-8 conversion
used in jsQR. In this case we fill the text entry from the bytes
and we set its encoding to "latin1".

See also:
https://github.com/cozmo/jsQR/issues/129

This solves issue:
https://github.com/brianwdavis/quadrangle/issues/9